requirements.txt
Django==5.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
Pillow==10.1.0
redis==5.0.1
celery==5.3.6
